Create  Sharing is Caring.py
A and B have a box of chocolate. B takes X chocolates from the box and A gets the remaining chocolates. Now if x>y, B puts y number of chocolates into the box. However if y>x, A puts x number of chocolates back into the box. Once the chocolates are put back they repeat the process with the remaining chocolates they have. This continues untill one of them does not have chocolates or both are left with equal number of chocolates. Write a program to find out Number of choclates A and B will be left with?

Input Format

The input two integers N and X seperated by a single white space. N - number of chocolates in the box X - number of chocolates B takes from the box initially

Constraints

N>=2 0

Output Format

The output contains the number of choclates they both are left with at the end.

Sample Input 0

10 4

Sample Output 0

2